IA-DEVELOPER-N2/U04-Desarrollo-Apps-Practicas-Clasificacion/02-Laboratorios/Lab-4.1-Medical-Images/requirements.txt
# Laboratorio 4.1 - Clasificación de Imágenes Médicas
# Dependencias principales
tensorflow==2.15.0
opencv-python==4.9.0
albumentations==1.3.0
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.2
seaborn==0.12.2
scikit-learn==1.3.0

# API y servidor
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pillow==10.0.1

# Utilidades
tqdm==4.66.1
python-dotenv==1.0.0
requests==2.31.0
